requirements.txt
pandas==2.0.3
sqlalchemy==2.0.20
psycopg2-binary==2.9.6
python-chrono==0.6.0
apache-airflow==2.7.2
pendulum==2.1.2